supera-tech-test/products_api/requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
kafka-python==2.0.2
pytz==2022.7.1
sqlparse==0.4.3
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
kafka-python==2.0.2
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
